Put all application-specific JS in one file
I don't think there will be enough code to warrant the cruft of
RequireJS, but it seemed nice to be able to load the code in either
the browser or Node, which would mean lots of repeated boilerplate.

For now, just put everything in one file and factor it out again if
it needed.