Add search
Add search using lunr.js.  Right now the search is just on the
provider name, but it would be easy to add other fields to the
index.

http://blog.apps.chicagotribune.com/2014/02/17/nuts-and-bolts-of-event-listings/
was extremely helpful.

Add a method to the Providers collection to reset filters and
trigger a unified ``filter`` event on either search or facet in
addition to explicit ``search`` and ``facet`` events.