Remove off-canvas behavior
We're moving the provider list and the filters to separate views,
so we don't want/need the off-canvas filter sidebar.

Keep the container elements in case we change our minds, but remove
the CSS that we needed to make the Bootstrap navbar work with the
off-canvas elements.  This makes the markup and styles for the
navbar pretty much the [Bootstrap
default](http://getbootstrap.com/components/#navbar).

Addresses https://pm.theworkdept.com/issues/3780