Don't select all available options
Thinking more about this, I think the behavior where all available
options are shown as selected is confusing.  I think it's better
to show just the items that were explicitly clicked as selected.

Also factor out the code that enables/disables select elements
based on the number of elements and the code that sizes the
select elements based on the number of elements.

Debounce the select change event handler to allow for easier
selection of multiple events in desktop browsers.

Addresses https://pm.theworkdept.com/issues/3917